+++ test/loop-batch.js
@@ -29,10 +29,11 @@ describe('Workflow - Loop Execution w/ Batch', function () {
         })
         .activity('activity1')
         .loop('loop1', ['activity1'], fragment, function () {
-          return ['test1', 'test2', 'test3'];
+          return ['test1', 'test2', 'test3', 'test4', 'test5'];
         }, {
           batchDelay: 1,
-          itemsPerBatch: 1
+          itemsPerBatch: 3,
+          maxOutstanding: 2
         })
         .activity('activity2', ['loop1']);
 
@@ -68,6 +69,10 @@ describe('Workflow - Loop Execution w/ Batch', function () {
     expect(events.results('loop1-1-activity4')).to.deep.equal({ activity4: 'Activity 4 output' });
     expect(events.results('loop1-2-activity3')).to.deep.equal({ activity3: 'Activity 3 output' });
     expect(events.results('loop1-2-activity4')).to.deep.equal({ activity4: 'Activity 4 output' });
+    expect(events.results('loop1-3-activity3')).to.deep.equal({ activity3: 'Activity 3 output' });
+    expect(events.results('loop1-3-activity4')).to.deep.equal({ activity4: 'Activity 4 output' });
+    expect(events.results('loop1-4-activity3')).to.deep.equal({ activity3: 'Activity 3 output' });
+    expect(events.results('loop1-4-activity4')).to.deep.equal({ activity4: 'Activity 4 output' });
     expect(events.results('activity2')).to.deep.equal({ activity2: 'Activity 2 output' });
   });